Crate starlark_syntax

Starlark AST.

Re-exports

pub use crate::error::Error;

pub use crate::error::ErrorKind;

pub use crate::error::StarlarkResultExt;

Modules

call_stack

Starlark call stack.

codemap

A data structure for tracking source positions in language implementations The CodeMap tracks all source files and maps positions within them to linear indexes as if all source files were concatenated. This allows a source position to be represented by a small 32-bit Pos indexing into the CodeMap, under the assumption that the total amount of parsed source code will not exceed 4GiB. The CodeMap can look up the source file, line, and column of a Pos or Span, as well as provide source code snippets for error reporting.

convert_indices

diagnostic

dialect

dot_format_parser

error

eval_exception

fast_string

Our string operations (indexing) are O(n) because of our current representation. There are plans afoot to change that, but in the meantime let’s use fast algorithms to make up some of the difference.

frame

golden_test_template

lexer

slice_vec_ext

span_display

syntax

The AST of Starlark as AstModule, along with a parse function.

Macros

function_error

internal_error

Internal error of starlark.

other_error

value_error

Type Aliases

Result