starlark_syntax

Module codemap

A data structure for tracking source positions in language implementations The CodeMap tracks all source files and maps positions within them to linear indexes as if all source files were concatenated. This allows a source position to be represented by a small 32-bit Pos indexing into the CodeMap, under the assumption that the total amount of parsed source code will not exceed 4GiB. The CodeMap can look up the source file, line, and column of a Pos or Span, as well as provide source code snippets for error reporting.

Structs

CodeMap

A data structure recording a source code file for position lookup.

CodeMapId

CodeMaps

Multiple CodeMap.

FileSpan

A file, and a line and column range within it.

FileSpanRef

A file, and a line and column range within it.

NativeCodeMap

“Codemap” for .rs files.

Pos

A small, Copy, value representing a position in a CodeMap’s file.

ResolvedFileLine

File and line number.

ResolvedFileSpan

File name and line and column pairs for a span.

ResolvedPos

All are 0-based, but print out with 1-based.

ResolvedSpan

The locations of values within a span. All are 0-based, but print out with 1-based.

Span

A range of text within a CodeMap.

Spanned

Associate a Span with a value of arbitrary type (e.g. an AST node).