starlark_syntax::error

Struct Error

pub struct Error(/* private fields */);

An error produced by starlark.

This error is composed of an error kind, together with some diagnostic information indicating where it occurred.

In order to prevent accidental conversions to anyhow::Error, this type intentionally does not implement std::error::Error. That should probably change in the future.

Implementations

impl Error

pub fn new_kind(kind: ErrorKind) -> Self

Create a new error

pub fn new_spanned(kind: ErrorKind, span: Span, codemap: &CodeMap) -> Self

Create a new error with a span

pub fn new_other(e: impl Into<Error>) -> Self

Create a new error with no diagnostic and of kind ErrorKind::Other

pub fn new_native(e: impl Into<Error>) -> Self

Create a new error with no diagnostic and of kind ErrorKind::Native

pub fn new_value(e: impl Into<Error>) -> Self

Create a new error with no diagnostic and of kind ErrorKind::Value

pub fn kind(&self) -> &ErrorKind

The kind of this error

pub fn into_kind(self) -> ErrorKind

Convert the error into the underlying kind

pub fn has_diagnostic(&self) -> bool

pub fn into_anyhow(self) -> Error

Convert this error into an anyhow::Error

pub fn without_diagnostic<'a>(&'a self) -> impl Debug + Display + 'a

Returns a value that can be used to format this error without including the diagnostic information

This is the same as kind, just a bit more explicit.

pub fn span(&self) -> Option<&FileSpan>

pub fn call_stack(&self) -> &CallStack

pub fn set_span(&mut self, span: Span, codemap: &CodeMap)

Set the span, unless it’s already been set.

pub fn set_call_stack(&mut self, call_stack: impl FnOnce() -> CallStack)

Set the call_stack field, unless it’s already been set.

pub fn eprint(&self)

Print an error to the stderr stream. If the error has diagnostic information it will use color-codes when printing.

Note that this function doesn’t print any context information if the error is a diagnostic, so you might prefer to use eprintln!("{:#}"), err) if you suspect there is useful context (although you won’t get pretty colors).

pub fn into_internal_error(self) -> Error

Change error kind to internal error.