starlark_syntax

Module fast_string

Our string operations (indexing) are O(n) because of our current representation. There are plans afoot to change that, but in the meantime let’s use fast algorithms to make up some of the difference.

Structs

CharIndex

Index of a char in a string. This is different from string byte offset.

StrIndices

Result of applying start and end to a string.

Functions

at

Find the character at position i.

contains

convert_str_indices

Convert common start and end arguments of str functions like str.find.

count_matches

Find the number of times a needle occurs within a string, non-overlapping.

count_matches_byte

Find the number of times a needle byte occurs within a string. If the needle represents a complete character, this will be equivalent to doing search for that character in the string.

len

Find the length of the string in characters. If the length matches the length in bytes, the string must be 7bit ASCII.

split_at

Split the string at given char offset. None if offset is out of bounds.